+++ TelegramGroupsAdmin.Telegram/Services/Moderation/Actions/WarnHandler.cs
@@ -33,7 +33,7 @@ public class WarnHandler : IWarnHandler
         Actor executor,
         string? reason,
         long chatId,
-        long? messageId = null,
+        int? messageId = null,
         CancellationToken cancellationToken = default)
     {
         _logger.LogDebug(
